[cute] Remove dead code and de-duplicate AST-pass helpers
Cleanup pass over the CuTe backend (helion/_compiler/cute/). All changes
are behavior-preserving: dead-symbol deletions, write-only field/param
removals, and de-duplication of byte-identical helpers.

Dead code removed (verified zero references, incl. the codegen string
import map):
- reduce_helpers: the _cute_pre_vec_fold_fp32 family (never registered)
- cute_mma: _is_zero_acc_node, _tcgen05_epilogue_dest_expr
- online_to_3pass: _is_full_init, _matches_subscript
- hoist_loop_invariant_recip: _expr_text
- pipeline_inner_loads: _format_step_expr
- fuse_two_pass_loads: unreachable _CuteFuseTwoPassLoads.visit_FunctionDef
  and its vestigial ast.NodeTransformer base

Write-only fields / unused params removed:
- mma_support: 6 diagnostic fields on CuteMmaSupport; _probe_* now return bool
- layout: LayoutConstraint.required + its always-false branch
- indexing: CuteAffineRangeIndex.start / .length
- layout_rules: preferred_constraint_for_node's unused graph_info param
- cute_epilogue: _AuxiliaryTensorStep.forward_form (direction is baked into
  op_template)

De-duplication:
- Extracted the byte-identical _NameRefCollector / _names_read /
  _assignment_lhs_name into a new shared _ast_pass_utils.py, imported by the
  4 AST-pass modules. The genuinely divergent _names_written and
  _is_constexpr_v_loop are intentionally left per-module.

Net ~-230 lines. Verified: ./lint.sh clean;
HELION_BACKEND=cute pytest test/test_cute_*.py -> 604 passed, 0 failed.

stack-info: PR: https://github.com/pytorch/helion/pull/2724, branch: oulgen/stack/349
