+++ helion/_compiler/cute/mma_support.py
@@ -7,16 +7,10 @@ import torch
 
 @dataclasses.dataclass(frozen=True)
 class CuteMmaSupport:
-    device_name: str | None
-    capability: tuple[int, int] | None
-    cutlass_arch: str | None
     universal: bool
     warp_f16bf16: bool
     warpgroup_f16bf16: bool
     tcgen05_f16bf16: bool
-    warp_error: str | None = None
-    warpgroup_error: str | None = None
-    tcgen05_error: str | None = None
 
     @property
     def supported_impls(self) -> tuple[str, ...]:
@@ -47,18 +41,18 @@ def _current_cutlass_arch_name() -> str | None:
         return None
 
 
-def _probe_warp_f16bf16() -> tuple[bool, str | None]:
+def _probe_warp_f16bf16() -> bool:
     try:
         import cutlass
         from cutlass.cute.nvgpu import warp
 
         warp.MmaF16BF16Op(cutlass.Float16, cutlass.Float32, (16, 8, 16))
-        return True, None
-    except Exception as exc:
-        return False, f"{type(exc).__name__}: {exc}"
+        return True
+    except Exception:
+        return False
 
 
-def _probe_warpgroup_f16bf16() -> tuple[bool, str | None]:
+def _probe_warpgroup_f16bf16() -> bool:
     try:
         import cutlass
         from cutlass.cute.nvgpu import OperandMajorMode
@@ -72,12 +66,12 @@ def _probe_warpgroup_f16bf16() -> tuple[bool, str | None]:
             OperandMajorMode.K,
             OperandMajorMode.K,
         )
-        return True, None
-    except Exception as exc:
-        return False, f"{type(exc).__name__}: {exc}"
+        return True
+    except Exception:
+        return False
 
 
-def _probe_tcgen05_f16bf16() -> tuple[bool, str | None]:
+def _probe_tcgen05_f16bf16() -> bool:
     try:
         import cutlass
         from cutlass.cute.nvgpu import OperandMajorMode
@@ -92,46 +86,28 @@ def _probe_tcgen05_f16bf16() -> tuple[bool, str | None]:
             OperandMajorMode.K,
             OperandMajorMode.K,
         )
-        return True, None
-    except Exception as exc:
-        return False, f"{type(exc).__name__}: {exc}"
+        return True
+    except Exception:
+        return False
 
 
 def get_cute_mma_support() -> CuteMmaSupport:
     device = _current_cuda_device()
     if device is None:
         return CuteMmaSupport(
-            device_name=None,
-            capability=None,
-            cutlass_arch=None,
             universal=False,
             warp_f16bf16=False,
             warpgroup_f16bf16=False,
             tcgen05_f16bf16=False,
-            warp_error="CUDA unavailable",
-            warpgroup_error="CUDA unavailable",
-            tcgen05_error="CUDA unavailable",
         )
 
-    device_name = torch.cuda.get_device_name(device)
-    capability = torch.cuda.get_device_capability(device)
     cutlass_arch = _current_cutlass_arch_name()
 
     # The universal atom is the only lowering Helion currently wires up end-to-end.
     universal = cutlass_arch is not None
-    warp_ok, warp_error = _probe_warp_f16bf16()
-    warpgroup_ok, warpgroup_error = _probe_warpgroup_f16bf16()
-    tcgen05_ok, tcgen05_error = _probe_tcgen05_f16bf16()
-
     return CuteMmaSupport(
-        device_name=device_name,
-        capability=capability,
-        cutlass_arch=cutlass_arch,
         universal=universal,
-        warp_f16bf16=warp_ok,
-        warpgroup_f16bf16=warpgroup_ok,
-        tcgen05_f16bf16=tcgen05_ok,
-        warp_error=warp_error,
-        warpgroup_error=warpgroup_error,
-        tcgen05_error=tcgen05_error,
+        warp_f16bf16=_probe_warp_f16bf16(),
+        warpgroup_f16bf16=_probe_warpgroup_f16bf16(),
+        tcgen05_f16bf16=_probe_tcgen05_f16bf16(),
     )
